Use TS, abstraction
