+++ package.json
@@ -12,13 +12,13 @@
     "next": "^11.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "typescript": "^4.4.2",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "export": "next export",
-    "deploy": "rimraf node_modules/.cache && next build && next export && xcopy CNAME ./out && gh-pages -d out"
+    "export": "next export"
   }
 }
