--- pages/404.js
@@ -1,16 +1,0 @@
-// React
-import React from "react";
-
-import VerticalText from "../components/VerticalText";
-
-const PageNotFound = () => {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <VerticalText bold text="Not found" />
-      </header>
-    </div>
-  );
-};
-
-export default PageNotFound;
