+++ pages/404.tsx
@@ -1,0 +1,11 @@
+import VerticalText from "../components/VerticalText";
+
+export default function PageNotFound() {
+    return (
+        <div className="App">
+            <header className="App-header">
+                <VerticalText bold text="Not found"/>
+            </header>
+        </div>
+    );
+};
