+++ pages/_app.tsx
@@ -1,3 +1,4 @@
+import {AppProps} from "next/app";
 import { ThemeProvider } from "@material-ui/core/styles";
 import themeFile from "../util/theme";
 
@@ -5,7 +6,7 @@ import "../styles/App.css";
 import "../styles/index.css"
 
 
-function App({ Component, pageProps }) {
+export default function App({ Component, pageProps }: AppProps) {
     return (
         <div className="main-background" style={{ backgroundColor: "#282c34" }}>
             <ThemeProvider theme={themeFile}>
@@ -14,5 +15,3 @@ function App({ Component, pageProps }) {
         </div>
     )
 }
-
-export default App;
