+++ pages/stuff/rune.tsx
@@ -1,0 +1,9 @@
+import TranslatePage from "../../components/TranslatePage";
+
+// Rune translate table
+import rune from "../../data/rune.json";
+
+
+export default function RuneTextGenerator() {
+    return <TranslatePage name="Rune Text Generator" dictionary={rune} />
+};
