+++ pages/stuff/similar.tsx
@@ -1,0 +1,9 @@
+import TranslatePage from "../../components/TranslatePage";
+
+// Similar translate table
+import similar from "../../data/similar.json";
+
+
+export default function SimilarTextGenerator() {
+    return <TranslatePage name="Similar Text Generator" dictionary={similar} allowLowercase highlightUntranslated />
+};
